requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dj-database-url==0.5.0
Django==3.2.8
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
django-environ==0.8.1
django-heroku==0.0.0
django-timezone-field==4.2.1
gunicorn==20.1.0
idna==3.3
kombu==5.1.0
Pillow==8.4.0
prompt-toolkit==3.0.21
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
